requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.1.1
faiss-cpu==1.8.0
